# ml-close-loop-be/openapi.yaml
openapi: 3.1.0
info:
  title: DEFNEX MLOps Backend
  version: 0.1.0
paths:
...
  /api/v1/models/{model_id}/versions/{version}/decisions:
    post:
      tags:
      - Decisions
      summary: Create Decision
      operationId: create_decision_api_v1_models__model_id__versions__version__decisions_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      - name: version
        in: path
        required: true
        schema:
          type: integer
          title: Version
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/models/{model_id}/rollback:
    post:
      tags:
      - Decisions
      summary: Rollback Model
      operationId: rollback_model_api_v1_models__model_id__rollback_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/models/{model_id}/versions/{version}/deploy-staging:
    post:
      tags:
      - Decisions
      summary: Deploy To Staging Endpoint
      description: 'Ladder step 1 (issue #69 AC 1), PRD §16.2: deploy an EVALUATED
        candidate to staging without

        touching the production pointer. Records the STAGING decision; smoke test
        runs automatically as

        part of the deploy (PRD §38.4).'
      operationId: deploy_to_staging_endpoint_api_v1_models__model_id__versions__version__deploy_staging_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      - name: version
        in: path
        required: true
        schema:
          type: integer
          title: Version
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LadderActionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/models/{model_id}/versions/{version}/validate-staging:
    post:
      tags:
      - Decisions
      summary: Validate Staging Endpoint
      description: 'Ladder step 2 (issue #69 AC 2/3): the human quality-gate approval
        that the staged candidate

        passed integration validation. Only a STAGING version may be validated, which
        structurally

        blocks production promotion until staging happened. Records the VALIDATED
        decision with the

        frozen evaluation snapshot as the gate result (PRD §8.4).'
      operationId: validate_staging_endpoint_api_v1_models__model_id__versions__version__validate_staging_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      - name: version
        in: path
        required: true
        schema:
          type: integer
          title: Version
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LadderActionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/models/{model_id}/versions/{version}/promote-production:
    post:
      tags:
      - Decisions
      summary: Promote Production Endpoint
      description: 'Ladder step 3 (issue #69 AC 2/4/5): the authorized controlled
        promotion of a VALIDATED

        (or legacy PROMOTED) candidate to the production pointer. Records the PRODUCTION
        decision;

        the registry terminal status stays DEPLOYED (single source of truth for production).'
      operationId: promote_production_endpoint_api_v1_models__model_id__versions__version__promote_production_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      - name: version
        in: path
        required: true
        schema:
          type: integer
          title: Version
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LadderActionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/deployments/{deployment_id}/rollback:
    post:
      tags:
      - Decisions
      summary: Rollback Deployment
      description: 'Issue #70, PRD §14.4 "Rollback": restore the previous immutable
        version for a deployment

        target, discovered by the deployment row (POST .../deployments/{id}/rollback,
        PRD §25). The

        version is located via the deployment''s recorded (model_id, model_version),
        the pointer is

        moved to the previous version, and the ROLLBACK decision records which environment
        was rolled

        back. The previous version is never deleted - artifacts are immutable (artifact_storage).'
      operationId: rollback_deployment_api_v1_deployments__deployment_id__rollback_post
      parameters:
      - name: deployment_id
        in: path
        required: true
        schema:
          type: string
          title: Deployment Id
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LadderActionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DecisionCreateRequest:
      properties:
        decision:
          type: string
          enum:
          - PROMOTED
          - REJECTED
          title: Decision
        decided_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Decided By
        rationale:
          type: string
          title: Rationale
      type: object
      required:
      - decision
      - rationale
      title: DecisionCreateRequest
      description: 'Request body for POST .../decisions (openapi.yaml DecisionCreateRequest).
        Restricted to

        PROMOTED|REJECTED - ROLLBACK is only ever produced via POST /models/{model_id}/rollback
        (US-017).'
    DecisionRecord:
      properties:
        decision_id:
          type: string
          title: Decision Id
        model_id:
          type: string
          title: Model Id
        version:
          type: integer
          title: Version
        decision:
          type: string
          enum:
          - PROMOTED
          - REJECTED
          - ROLLBACK
          - STAGING
          - VALIDATED
          - PRODUCTION
          title: Decision
        decided_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Decided By
        decided_at:
          type: string
          format: date-time
          title: Decided At
        evidence_snapshot:
          anyOf:
          - $ref: '#/components/schemas/EvaluationObject'
          - type: 'null'
        eval_set_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Eval Set Id
        eval_set_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Eval Set Version
        rationale:
          type: string
          title: Rationale
        rollback_of_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rollback Of Version
        dataset_license:
          anyOf:
          - type: string
          - type: 'null'
          title: Dataset License
        base_model_license:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Model License
        license_warning:
          anyOf:
          - type: string
          - type: 'null'
          title: License Warning
      type: object
      required:
      - decision_id
      - model_id
      - version
      - decision
      - decided_at
      - rationale
      title: DecisionRecord
      description: 'WBS 3.3 §8 decision record (openapi.yaml DecisionRecord) - covers
        promotion, rejection,

        rollback, and the staging/production ladder steps with one schema. `dataset_license`/
        `base_model_license`/`license_warning` (issue #134) surface the training dataset and
        base model licenses and an explicit, human-facing (never auto-blocking) warning when a
        non-commercial dataset license is heading toward production.'
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
      - code
      - message
      title: ErrorDetail
      description: openapi.yaml Error.error (PRD §18 standard error envelope).
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorResponse
      description: 'openapi.yaml Error — `{"error": {"code": ..., "message": ...}}`.'
    EvalLossTrend:
      properties:
        previous_version_eval_loss:
          anyOf:
          - type: number
          - type: 'null'
          title: Previous Version Eval Loss
        this_version_eval_loss:
          type: number
          title: This Version Eval Loss
      type: object
      required:
      - this_version_eval_loss
      title: EvalLossTrend
...
    EvaluationObject:
      properties:
        eval_loss_trend:
          anyOf:
          - $ref: '#/components/schemas/EvalLossTrend'
          - type: 'null'
        qualitative_comparison:
          anyOf:
          - $ref: '#/components/schemas/QualitativeComparison'
          - type: 'null'
        general_domain_regression_check:
          anyOf:
          - $ref: '#/components/schemas/GeneralDomainRegressionCheck'
          - type: 'null'
      type: object
      title: EvaluationObject
      description: The three-signal evaluation payload (openapi.yaml EvaluationObject).
...
    GeneralDomainRegressionCheck:
      properties:
        checked:
          type: boolean
          title: Checked
        regressions_found:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Regressions Found
          default: []
      type: object
      required:
      - checked
      title: GeneralDomainRegressionCheck
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LadderActionRequest:
      properties:
        decided_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Decided By
        rationale:
          type: string
          title: Rationale
      type: object
      required:
      - rationale
      title: LadderActionRequest
      description: 'Request body for the staging/production ladder steps and the deployment-level
        rollback

        (openapi.yaml LadderActionRequest). The version being acted on is in the URL;
        the body carries

        only the audit fields (who approved, why), matching `DecisionCreateRequest.decided_by`.'
...
    QualitativeComparison:
      properties:
        question_table_version:
          type: integer
          title: Question Table Version
        wins:
          type: integer
          title: Wins
        losses:
          type: integer
          title: Losses
        ties:
          type: integer
          title: Ties
        total:
          type: integer
          title: Total
      type: object
      required:
      - question_table_version
      - wins
      - losses
      - ties
      - total
      title: QualitativeComparison
...
    RollbackRequest:
      properties:
        rollback_of_version:
          type: integer
          title: Rollback Of Version
        decided_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Decided By
        rationale:
          type: string
          title: Rationale
      type: object
      required:
      - rollback_of_version
      - rationale
      title: RollbackRequest
      description: Request body for POST /models/{model_id}/rollback (openapi.yaml
        RollbackRequest).
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...